Reached to Ingredients App

+++ demo/src/main/java/com/example/demo/domain/ingredientdomain/IngredientProjection.java
@@ -1,11 +1,12 @@
 package com.example.demo.domain.ingredientdomain;
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface IngredientProjection {
-    
+
     UUID getId();
 
     String getName();
 
-    Double getPrice();
+    BigDecimal getPrice();
 }
